SECURITY [XSS-Injection]: Properly escape i18n message calls that had no output mode
These messages were being passed through Html::rawElement(), which
explicitly has no escaping built-in. These messages also had no
explicit output mode(), which allowed XSS injections.

It's really better to use Html::element() instead, and in the future there
should/could be a migration to that, but we are migrating to Mustache in
the future.

Bug: T265440
Change-Id: I4e3152e988706fdfc318cf0d17b594ceac15c170